ENT-2291: Create adapter methods for DTOs implementing *Info Interfaces  - Added a class InfoAdapter to provide implementation    of info interfaces.  - Removed the translators and its test suites.    1. UserDTOTranslator    2. RoleDTOTranslator    3. PermissionBluePrintDTOTranslator
